+++ sampledb/logic/actions.py
@@ -245,7 +245,9 @@ def create_action(
         instrument_id: typing.Optional[int] = None,
         user_id: typing.Optional[int] = None,
         description_is_markdown: bool = False,
-        is_hidden: bool = False
+        is_hidden: bool = False,
+        short_description: str = '',
+        short_description_is_markdown: bool = False
 ) -> Action:
     """
     Creates a new action with the given type, name, description and schema. If
@@ -260,6 +262,9 @@ def create_action(
     :param user_id: None or the ID of an existing user
     :param description_is_markdown: whether the description contains Markdown
     :param is_hidden: None or whether or not the action should be hidden
+    :param short_description: the new (possibly empty) short description
+    :param short_description_is_markdown: whether the short description
+        contains Markdown
     :return: the created action
     :raise errors.ActionTypeDoesNotExistError: when no action type with the
         given action type ID exists
@@ -288,7 +293,9 @@ def create_action(
         is_hidden=is_hidden,
         schema=schema,
         instrument_id=instrument_id,
-        user_id=user_id
+        user_id=user_id,
+        short_description=short_description,
+        short_description_is_markdown=short_description_is_markdown
     )
     db.session.add(action)
     db.session.commit()
@@ -334,7 +341,9 @@ def update_action(
         description: str,
         schema: dict,
         description_is_markdown: bool = False,
-        is_hidden: typing.Optional[bool] = None
+        is_hidden: typing.Optional[bool] = None,
+        short_description: str = '',
+        short_description_is_markdown: bool = False
 ) -> None:
     """
     Updates the action with the given action ID, setting its name, description and schema.
@@ -345,6 +354,9 @@ def update_action(
     :param schema: the new schema for objects created using this action
     :param description_is_markdown: whether the description contains Markdown
     :param is_hidden: None or whether or not the action should be hidden
+    :param short_description: the new (possibly empty) short description
+    :param short_description_is_markdown: whether the short description
+        contains Markdown
     :raise errors.SchemaValidationError: when the schema is invalid
     :raise errors.InstrumentDoesNotExistError: when instrument_id is not None
         and no instrument with the given instrument ID exists
@@ -357,6 +369,8 @@ def update_action(
     action.description = description
     action.description_is_markdown = description_is_markdown
     action.schema = schema
+    action.short_description = short_description
+    action.short_description_is_markdown = short_description_is_markdown
     if is_hidden is not None:
         action.is_hidden = is_hidden
     db.session.add(action)